backend/pyproject.toml
[tool.poetry]
name = "ai-digital-workforce-backend"
version = "0.1.0"
description = "Backend API for AI Digital Workforce - Multi-agent AI collaboration platform"
authors = ["AutoMate <contact@automate.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
sqlalchemy = "^2.0.23"
alembic = "^1.13.1"
python-socketio = "^5.10.0"
python-multipart = "^0.0.6"
openai = "^1.3.0"
langchain = "^0.0.340"
langgraph = "^0.0.20"
tavily-python = "^0.3.0"
python-dotenv = "^1.0.0"
httpx = "^0.25.2"
aiofiles = "^23.2.1"
jinja2 = "^3.1.2"
passlib = "^1.7.4"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
black = "^23.11.0"
flake8 = "^6.1.0"
mypy = "^1.7.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true